restructure reducers
