--- src/reducers/InitialState.js
@@ -1,5 +1,0 @@
-export const initialState = {
-  path: []
-  artistInfo: {}
-  pathSuccess: false
-}
