--- src/reducers/RootReducer.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import pathReducer from './PathReducer';
-
-export default combineReducers({
- pathReducer
-});
