+++ src/reducers/paths/actions.js
@@ -1,4 +1,4 @@
-import ArtistsApi from '../api/artists';
+import ArtistsApi from '../../api/artists';
 
 export const fetchPathsSuccess = (paths) => dispatch => {
  dispatch({
@@ -13,13 +13,6 @@ export const fetchPathsError = () => dispatch => {
  })
 }
 
-export const fetchArtistInfoSuccess = (artistInfo) => dispatch => {
- dispatch({
-  type: 'FETCH_ARTIST_INFO_SUCCESS',
-  payload: artistInfo
- })
-}
-
 export const fetchPaths = (dispatch) => {
   return async (dispatch) => {
     try {
@@ -28,18 +21,6 @@ export const fetchPaths = (dispatch) => {
       dispatch(fetchPathsSuccess(responseJson))
     } catch (error) {
       dispatch(fetchPathsError())
-    }
-  };
-}
-
-export const fetchArtistInfo = (dispatch) => {
-  return async (dispatch) => {
-    try {
-      const response = await ArtistsApi.fetchArtistInfo(1)
-      const responseJson = await response.json()
-      dispatch(fetchArtistInfoSuccess(responseJson))
-    } catch (error) {
-      console.log('An error occurred.', error)
     }
   };
 }
